Organize code in RaftStateContext and add postAppendEntries method for post execution operations
