+++ servlet/src/main/java/com/springRaft/servlet/consensusModule/Candidate.java
@@ -206,25 +206,14 @@ public class Candidate extends RaftStateContext implements RaftState {
     /* --------------------------------------------------- */
 
     /**
-     * A method that encapsulates replicated code, and has the function of setting
-     * the reply for the received AppendEntries.
+     * Implementation of the postAppendEntries abstract method in parent class.
+     * This method contains the behaviour to execute after invoking appendEntries method.
      *
      * @param appendEntries The received AppendEntries communication.
-     * @param reply AppendEntriesReply object, to send as response to the leader.
      * */
-    protected void setAppendEntriesReply(AppendEntries appendEntries, AppendEntriesReply reply) {
+    protected void postAppendEntries(AppendEntries appendEntries) {
 
         this.cleanBeforeTransit();
-
-        // reply with the current term
-        reply.setTerm(appendEntries.getTerm());
-
-        // check reply's success based on prevLogIndex and prevLogTerm
-        // reply.setSuccess()
-        // ...
-        // ...
-        // this need to be changed
-        reply.setSuccess(true);
 
         // transit to follower state
         this.transitionManager.setNewFollowerState();
