+++ servlet/src/main/java/com/springRaft/servlet/consensusModule/Follower.java
@@ -3,9 +3,7 @@ package com.springRaft.servlet.consensusModule;
 import com.springRaft.servlet.communication.message.*;
 import com.springRaft.servlet.communication.outbound.OutboundManager;
 import com.springRaft.servlet.config.RaftProperties;
-import com.springRaft.servlet.persistence.log.Entry;
 import com.springRaft.servlet.persistence.log.LogService;
-import com.springRaft.servlet.persistence.log.LogState;
 import com.springRaft.servlet.persistence.state.StateService;
 import com.springRaft.servlet.util.Pair;
 import org.slf4j.Logger;
@@ -165,89 +163,21 @@ public class Follower extends RaftStateContext implements RaftState {
     /* --------------------------------------------------- */
 
     /**
-     * A method that encapsulates replicated code, and has the function of setting
-     * the reply for the received AppendEntries.
+     * Implementation of the postAppendEntries abstract method in parent class.
+     * This method contains the behaviour to execute after invoking appendEntries method.
      *
      * @param appendEntries The received AppendEntries communication.
-     * @param reply AppendEntriesReply object, to send as response to the leader.
      * */
-    protected void setAppendEntriesReply(AppendEntries appendEntries, AppendEntriesReply reply) {
+    @Override
+    protected void postAppendEntries(AppendEntries appendEntries) {
 
         this.leaderId = appendEntries.getLeaderId();
 
         // remove the scheduled task
         this.transitionManager.cancelScheduledTask(this.scheduledFuture);
 
-        // reply with the current term
-        reply.setTerm(appendEntries.getTerm());
-
-        // check reply's success based on prevLogIndex and prevLogTerm
-        Entry entry = this.logService.getEntryByIndex(appendEntries.getPrevLogIndex());
-        entry = entry == null ? new Entry((long) 0, (long) 0, null) : entry;
-
-        if(entry.getIndex() == (long) appendEntries.getPrevLogIndex()) {
-
-            if (entry.getTerm() == (long) appendEntries.getTerm()) {
-
-                reply.setSuccess(true);
-
-                this.applyAppendEntries(appendEntries);
-
-            } else {
-
-                reply.setSuccess(false);
-
-            }
-
-        } else {
-
-            reply.setSuccess(false);
-
-        }
-
         // set a new timeout, it's equivalent to transit to a new follower state
         this.setTimeout();
-
-    }
-
-    /**
-     * Method for insert new entries in log and update the committed values in log state.
-     *
-     * @param appendEntries The received AppendEntries communication.
-     * */
-    private void applyAppendEntries(AppendEntries appendEntries) {
-
-        if (appendEntries.getEntries().size() != 0) {
-
-            // delete all the following conflict entries
-            this.logService.deleteIndexesGreaterThan(appendEntries.getPrevLogIndex());
-
-            // insert new entry
-            Entry newEntry = new Entry(appendEntries.getTerm(), appendEntries.getEntries().get(0));
-            newEntry = this.logService.insertEntry(newEntry);
-
-            // update committed entries in LogState
-            LogState logState = this.logService.getState();
-            if (appendEntries.getLeaderCommit() > logState.getCommittedIndex()) {
-
-                if (newEntry.getIndex() <= appendEntries.getLeaderCommit()) {
-
-                    logState.setCommittedIndex(newEntry.getIndex());
-                    logState.setCommittedTerm(newEntry.getTerm());
-                    this.logService.saveState(logState);
-
-                } else {
-
-                    long index = appendEntries.getLeaderCommit();
-                    Entry entry = this.logService.getEntryByIndex(index);
-                    logState.setCommittedIndex(entry.getIndex());
-                    logState.setCommittedTerm(entry.getTerm());
-
-                }
-
-            }
-
-        }
 
     }
